Revert "chore(autoconf): Introducing methods to handle dataset URLs"
This reverts commit 0145e5543a4bfbb7f28e80aefea94b505b812e02.
